Compat fixes

+++ src/main/java/com/talhanation/recruits/entities/ai/navigation/RecruitPathNavigation.java
@@ -50,6 +50,6 @@ public class RecruitPathNavigation extends AsyncGroundPathNavigation {
     public boolean moveTo(double x, double y, double z, double speed) {
         this.recruit.setMaxFallDistance(1);
         ((RecruitsPathNodeEvaluator) this.nodeEvaluator).setTarget((int) x, (int) y, (int) z);
-        return this.moveTo(this.createPath(new BlockPos(x, y, z), 0), speed);
+        return this.moveTo(this.createPath(new BlockPos((int) x, (int) y, (int) z), 0), speed);
     }
 }
